Глава 3. Упр 5.

+++ code.cpp
@@ -6,7 +6,7 @@ using namespace std;
 //c++ -o code code.cpp
 int main () {
 	cout << "Введите через пробел val1 и val2: ";
-	int val1,val2;
+	double val1,val2;
 	cin >> val1 >> val2;
 	if (val1 == val2) cout << "val1 = val2\n";
 		else if (val1 > val2) cout << "val1 > val2\n";
